Check directories before deleting folder

+++ 0 Setup/Setup.cs
@@ -166,7 +166,9 @@ public partial class Setup : Node
 
     void DeleteDirectoryIfEmpty(string path)
     {
-        if (Directory.Exists(path) && Directory.GetFiles(path).Length == 0)
+        if (Directory.Exists(path) 
+            && Directory.GetFiles(path).Length == 0
+            && Directory.GetDirectories(path).Length == 0)
         {
             Directory.Delete(path);
         }
